DESIGN and INSTALL documentation
The DESIGN doc has some brief and general architecture information for people looking to see how all the parts work together.

The INSTALL documentation is a first version of what someone would need to do to have Rainmap running. It does not include information on configuring and using Memcached or RabbitMQ, but it does go through the process of getting a PostgreSQL, Apache, Nginx and Django working.