Do a gettimeofday to refresh nsock_tod at the top of nsock_pcap_open.
jah found that the first pcap read was always timing out in qscan.nse.
http://seclists.org/nmap-dev/2010/q3/542. According to the logs, the
read event was happening around a second before the script's first probe
was sent, even though the script does things in the opposite order. The
read event was already long timed out by the time the probe was sent.
This only happened to the first read.

What was happening was that nsock_tod, Nsock's cache of the
gettimeofday, was stale. It still had the value it was initialized with
when nsp_new (via open_nse) was called, even before port scanning was
done. nsock_tod is updated frequently in nsock_loop but not outside it.
The first read event was scheduled before nsock_loop was entered, so it
had the seconds-old timeout, but subsequent reads were fine.

I think we really should be updating nsock_tod when every event is
added. But I didn't find a good centralized place to do it;
nsp_add_event already gets an absolute timeout. So I took a cue from the
connect functions. Outside of nsock_loop, nsock_tod is only updated in
nsock_connect_*. I think the logic here is that you might wait a while
before starting a connection, but then subsequent events on the same iod
will come quickly. The same thing happens with nsock_pcap_open; once you
open it it's likely that you will start new events soon, so it's a good
place to update.