Use the M flag (multiline) when matching strings for highlighting.
This is required to make '^' match right after a newline, in addition to
the beginning of a string. Now that we do highlighting in batches of
several lines instead of one line at a time, this is necessary. For
example, "NSE:" at the beginning of a line was only being highlighted if
it was also at the beginning of a batch.