Added a new option: --save <file> which gives the user the ability to specify the filename of the restoration file. By default Ncrack saves this file at ~/.ncrack/.restore-date (as usual)
In addition, whenever the user has specified a service with a timeout (using
the -to timing option), then Ncrack will automatically save a restoration file.
Normally Ncrack only does that if the user sends a SIGTERM/SIGINT (normally by pressing ctrl+C), 
but the reasoning behind leaving a restoration file behind whenever there are services
that have been imposed a timeout by the user is that in that case there is a
big chance that the user might want to resume that session in the future (it
hasn't finished being cracked after all - the internal timeout timer was just
triggered as specified by the user through the -to option).

The above, in combination with the ability to give a specific filename to the
saved file, gives the user the flexibility to schedule his scans (through cron
for example) using -to <time> to specify the number of hours/minutes that Ncrack will
be running and then saving its current progress in a specific
filename. Then, the scheduler can continue running Ncrack by using --resume
<filename>. Ncrack will then run for another <time> resuming from where it
stopped the last time.