Added skeletton for the first (select) engine.
Each engine will export a clearly defined interface to nsock, but will implement
it using a specific IO events notification function.

This design allows to make use of the most efficient system available, even if
it is not portable, without breaking compatibility with other operating systems.
Engines implementations are isolated from each other and should therefore be
easily maintainable.

Adding support for a new function (like the linux epoll or BSD kqueue) should be
as easy as writing the corresponding engine.