Remove tempaddr.
I asked Colin to define tempaddr for better modularity of address
handling between connect and listen modes. The idea was to do this:

	tempaddr = command-line address
	if (listen_mode)
		srcaddr = tempaddr
	else if (connect_mode)
		targetaddr = tempaddr

But I think that approach needs more work to work correctly. Now the
logic is more like it was before:

	targetaddr = command-line address
	if (listen_mode)
		srcaddr = targetaddr

The problem I had with this is that it overloads the meaning of
targetaddr, but it's fine for now.