Improved performances by letting epoll return as many events as possible, regardless how many IODs we are dealing with.
The event array of struct epoll_event is not allocated on the stack anymore in
epoll_loop() but is dynamically allocated, and available in the
epoll_engine_info structure. The array is resized if the number of IODs
increases.

Also removed the tmp_active_iods list and the associated loop in
iterate_through_event_lists(). Move the processed IODs to the beginning of the
active_iods list instead and start iterating after them then.