Check if the specified vuln IDs reference different vulnerabilities that were added previously to the database, if so then print a debug message for each returned vulnerability.
So we do not ignore the new vulnerability entry since we can't know which
one is the correct one, the previous or the new one ?
This is not the job of this library, it's the job of scripts to make sure
that IDs correspond to the correct vulnerabilities.


e.g:

Script X do:
add({title = "Program vulnerability X", state = vulns.State.LIKELY_VULN,
    IDS={'CVE'='CVE-0001-0001','OSVDB'=0001})

Script Y do:
add({title = "Program vulnerability Y", state = vulns.State.VULN,
    IDS={'CVE'='CVE-0001-0001','OSVDB'=0002})

Script Z do:
add({title = "Program vulnerability Z", state = vulns.State.LIKELY_VULN,
    IDS={'CVE'='CVE-0001-0001','OSVDB'=0002})


In this case the scripts report different vulnerabilities but they
use the same IDs CVE or OSVDB-ID. We do not know which one is the
correct one, or which is the buggy one: IDs ? or vulnerability entry ?

Scenarios:
o  Script X adds the X vulnerability then script Y adds the Y vulnerability.
   The library will inspect the IDs of the Y vulnerability and detects that
   CVE ID reference the X vulnerability (added by X script) so it will
   update the info of the X vulnerability (e.g: 'LIKELY_VULNERABLE' to
   'VULNERABLE', etc) and it will also save the Y vulnerability in the
   database, and updates the OSVDB-ID=0002 to point to it (Y in the database).

   We choosed this design since there are vulnerability information that
   can be updated by other scripts. A first script can detects that the
   program seems 'LIKELY VULNERABLE' and a second script can confirm the
   vulnerability so we must change the state from 'LIKELY_VULNERABLE' to
   'VULNERABLE'.

   When the Script Z adds the Z vulnerability, the library will detect that
   the CVE ID and the OSVDB-ID point to two different vulnerabilities:
   CVE='CVE-0001-0001' points to the X vulnerability.
   OSVDB=0002   points to the Y vulnerability.
   In this case a debug message will be printed for each vulnerability
   (X and Y) to inform the user that there is a problem with vulnerability
   IDs. And the vulnerability Z will not be saved in the database
   (registry) since both of its IDs reference previously saved
   vulnerabilities, so the library will just update the X and Y
   information if necessary, and the Z vulnerability will be discarded.

   Conclusion: scripts must provide the correct IDs.


Later we'll do more review and tests on other scenarios and code paths.