Make FPResponse into a dumb container for bytes, not a subclass of FPPacket.
I want to defer the parsing of packets until after the
information-capturing phase, at the classification stage. Recovering the
byte buffer from PacketElements requires perfect round-tripping, which
is unnecessary. This also make memory management simpler: it would be
necessary to define copy constructors and all that in order to assign a
chain of PacketElement.