Use the raw logistic values from predict_values.
This is as opposed to the probability values from predict_probability.
Say that the raw activation values for three classes were
	1.5	0.5	-2.0
The logistic of each of these is
	0.81757	0.62245	0.11920
So now we would report the probabilities 82%, 62%, and 12%. What
predict_probability does is calculate each of these values' fraction out
of the sum of all values:
	0.52434 = 0.81757 / (0.81757 + 0.62245 + 0.11920)
	0.39920 = 0.62245 / (0.81757 + 0.62245 + 0.11920)
	0.07644 = 0.11920 / (0.81757 + 0.62245 + 0.11920)
So it would report 52%, 40%, and 8%.

An instance where this matters is if values have little relative
difference but are far from zero. For example, we would want to report
the probabilities for
	-10	-10	-10
as much lower than for
	-1	-1	-1
though predict_probability would report 33.3% for all classes in both of
these cases.