Change META to EXTRA.
I think that the SCAN line is already "meta". The "extra" is data that
is outside of the packet contents. I intened to use this, at least, to
store a flag indicating whether the response to a Node Information query
is a DNS name rather than IP addresses, after that information is
censored for submission.