$ nse_remove_require $ nse_check_globals
nse_remove_require is not in version control yet. It basically removes
all calls to the require function so nse_check_globals can add correct calls.

nse_check_globals has been overhauled to patch code to require each
external module needed. This is necessary because Lua modules are no longer
put in the global environment via the module function (we have our own module
function now; the other is deprecated but still available by default). Any
module used by code (including scripts) must actually be required and stored in
a local so it can be accessed.

In a previous revision I added a stdnse.module function which mimics the behavior
of the Lua 5.1 module function in all ways except it does not store the module
in the global environment. There are two reasons. One: this helps with adopting
the new Lua practice of libraries only being available through require.
Second: code which does not require a library it uses relies on that
library being loaded through some other means. Sometimes another script or
another library requires the library on your behalf. Or, NSE loads the library
before running any script code (e.g. the 'bin' library). This is a bug waiting to
happen as you cannot predict in what order scripts will run for a user. So, I have
gone to the trouble of also making nse_check_globals localize and require every
library (that we know about, in nselib or Lua) that code uses.

nse_check_globals still tells the developer about "bad" global accesses as it
did before in addition to this new behavior (which can be turned off).

The new nse_check_globals and nse_remove_require will be available in a future
revision to the private repository. There's still some polishing to do.