Structured output in place
Scripts can still return just a string, which will be used in the
`//script@output` attribute and as the CDATA contents of the script
element. If, however, they return a table, that table will be converted
to text using a similar method to value_dump from nse_debug.cc, and to
XML in a hierarchical fashion.