Move formatting code from C to Lua
Recursive text formatting of script return has been moved from C++
functions in nse_main.cc to Lua by requiring the use of the __tostring
metatable key. If a script returns a table with __tostring set, then it
is converted to a string using that method. If not, stdnse.format_table
is used as a default. This has the same output as the C++ functions did,
and is documented in Luadoc in stdnse.lua