Do automatic formatting of script output tables.
There is a new Lua function in nse_main.lua that does this. It is found
under the key "NSE_FORMAT_TABLE" in the Lua registry. The output is
indented for each level of table nesting. Tables with a __tostring
metamethod are formatted using tostring. Table values that are keyed by
integers (i.e., those that are found by ipairs) are shown first,
followed by "key: value" lines for each non-number key.

For example, the table
  {delta = "0", date="2012-08-02T22:35:33Z"}
would be formatted as
  delta: 0
  date: 2012-08-02T22:35:33Z