Limit ventrilo-info to port 3784 (TCP or UDP).
Cf. r30460. We may want to run on more port numbers than just this one,
because the server can be configured to run on other ports. But we don't
want it to run for every open|filtered UDP port by default. We don't
ahve a good way to express anything in between.