Make it a non-fatal error when one of two listeners can't be opened.
Unless given a specific listen address, we open two separate listening
sockets, one for IPv4 and one for IPv6. It was previously a fatal error if we
failed to create either socket. Now it is fatal only when all potential
listening addresses fail.

David Millis discovered that the IPv6 listener failed on Windows XP without
IPv6 configured.
Ncat: socket: An address incompatible with the requested protocol was used.  QUITTING.
http://seclists.org/nmap-dev/2013/q3/96