Defer starting the next recv, if the previous recv didn't call to Nsock.
What I found was a stack overflow (or what would have been a stack
overflow if the Lua garbage collector hadn't slowed it down), when a
recv method returned without calling back to Nsock. For example,
	recv = function(self) return "hello\n" end
What happens is that Ncat starts another read once the previous read has
been processed. Imagine a do_read function that uses read_handler as an
Nsock handler. But because the Lua function doesn't necessarily call
Nsock, do_read can call read_handler directly if the function returns
rather than yields. If read_handler calls do_read to schedule the next
read, this leads to an infinite loop of each function calling the other.

To fix this, we check whether a coroutine has yielded to Nsock; if not,
then we don't start the next read until the next round of Nsock (which
we do by making the call after a short timer expires).