Properly inherit send and recv.
Previously I just had unknown methods looked up in the super table by
means of an __index metamethod. This wasn't quite right because the
"self" argument to the function was still that of the subtable. That is,
we were essentially doing this:
	recv = function(self, ...)
		return super.recv(self, ...)
	end
What we need to do instead is call the super method with super as
"self", like this:
	recv = function(self, ...)
		return super.recv(super, ...)
	end
Where this fell apart was when the recursion finally bottomed out at the
C implementations of send and recv. They were expecting a userdata, not
a Lua table, and threw an error.